+++ packages/react-code-editor/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.3.31](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-code-editor@4.3.30...@patternfly/react-code-editor@4.3.31) (2021-08-18)
+
+**Note:** Version bump only for package @patternfly/react-code-editor
+
+
+
+
+
+## [4.3.30](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-code-editor@4.3.29...@patternfly/react-code-editor@4.3.30) (2021-08-17)
+
+**Note:** Version bump only for package @patternfly/react-code-editor
+
+
+
+
+
+## [4.3.29](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-code-editor@4.3.28...@patternfly/react-code-editor@4.3.29) (2021-08-16)
+
+**Note:** Version bump only for package @patternfly/react-code-editor
+
+
+
+
+
+## [4.3.28](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-code-editor@4.3.27...@patternfly/react-code-editor@4.3.28) (2021-08-12)
+
+**Note:** Version bump only for package @patternfly/react-code-editor
+
+
+
+
+
 ## [4.3.27](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-code-editor@4.3.26...@patternfly/react-code-editor@4.3.27) (2021-08-11)
 
 **Note:** Version bump only for package @patternfly/react-code-editor