+++ packages/react-docs/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.21.1](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.21.0...@patternfly/react-docs@5.21.1) (2021-08-18)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+# [5.21.0](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.20.12...@patternfly/react-docs@5.21.0) (2021-08-17)
+
+
+### Features
+
+* **Navigation:** add horizontal subnav variant ([#6149](https://github.com/patternfly/patternfly-react/issues/6149)) ([1ec8488](https://github.com/patternfly/patternfly-react/commit/1ec8488a3056fc659ec83f04a6e793903ddd2e83))
+
+
+
+
+
+## [5.20.12](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.20.11...@patternfly/react-docs@5.20.12) (2021-08-16)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.20.11](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.20.10...@patternfly/react-docs@5.20.11) (2021-08-16)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.20.10](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.20.9...@patternfly/react-docs@5.20.10) (2021-08-12)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
 ## [5.20.9](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.20.8...@patternfly/react-docs@5.20.9) (2021-08-11)
 
 **Note:** Version bump only for package @patternfly/react-docs