+++ packages/react-inline-edit-extension/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.7.53](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.7.52...@patternfly/react-inline-edit-extension@4.7.53) (2021-08-18)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.7.52](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.7.51...@patternfly/react-inline-edit-extension@4.7.52) (2021-08-17)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.7.51](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.7.50...@patternfly/react-inline-edit-extension@4.7.51) (2021-08-16)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.7.50](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.7.49...@patternfly/react-inline-edit-extension@4.7.50) (2021-08-12)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
 ## [4.7.49](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.7.48...@patternfly/react-inline-edit-extension@4.7.49) (2021-08-11)
 
 **Note:** Version bump only for package @patternfly/react-inline-edit-extension