+++ packages/react-log-viewer/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.4.9](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-log-viewer@4.4.8...@patternfly/react-log-viewer@4.4.9) (2021-08-18)
+
+**Note:** Version bump only for package @patternfly/react-log-viewer
+
+
+
+
+
+## [4.4.8](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-log-viewer@4.4.7...@patternfly/react-log-viewer@4.4.8) (2021-08-17)
+
+**Note:** Version bump only for package @patternfly/react-log-viewer
+
+
+
+
+
+## [4.4.7](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-log-viewer@4.4.6...@patternfly/react-log-viewer@4.4.7) (2021-08-16)
+
+**Note:** Version bump only for package @patternfly/react-log-viewer
+
+
+
+
+
+## [4.4.6](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-log-viewer@4.4.5...@patternfly/react-log-viewer@4.4.6) (2021-08-12)
+
+**Note:** Version bump only for package @patternfly/react-log-viewer
+
+
+
+
+
 ## [4.4.5](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-log-viewer@4.4.4...@patternfly/react-log-viewer@4.4.5) (2021-08-11)
 
 **Note:** Version bump only for package @patternfly/react-log-viewer