+++ packages/react-virtualized-extension/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.9.20](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-virtualized-extension@4.9.19...@patternfly/react-virtualized-extension@4.9.20) (2021-08-18)
+
+**Note:** Version bump only for package @patternfly/react-virtualized-extension
+
+
+
+
+
+## [4.9.19](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-virtualized-extension@4.9.18...@patternfly/react-virtualized-extension@4.9.19) (2021-08-17)
+
+**Note:** Version bump only for package @patternfly/react-virtualized-extension
+
+
+
+
+
+## [4.9.18](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-virtualized-extension@4.9.17...@patternfly/react-virtualized-extension@4.9.18) (2021-08-16)
+
+**Note:** Version bump only for package @patternfly/react-virtualized-extension
+
+
+
+
+
+## [4.9.17](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-virtualized-extension@4.9.16...@patternfly/react-virtualized-extension@4.9.17) (2021-08-12)
+
+**Note:** Version bump only for package @patternfly/react-virtualized-extension
+
+
+
+
+
 ## [4.9.16](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-virtualized-extension@4.9.15...@patternfly/react-virtualized-extension@4.9.16) (2021-08-11)
 
 **Note:** Version bump only for package @patternfly/react-virtualized-extension